% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-cenariosena.r
\name{as.cenariosena}
\alias{as.cenariosena}
\title{Conversao Para \code{cenariosena}}
\usage{
as.cenariosena(dat)
}
\arguments{
\item{dat}{data.table para converter}
}
\value{
objeto da classe \code{cenariosena}
}
\description{
Converte um data.table em um objeto da classe \code{cenariosena}
}
\details{
\code{dat} deve ter a seguinte estrutura\tabular{ccccc}{
   anoref \tab bacia \tab cenario \tab indice \tab valor \cr
   AA \tab SIN \tab 1 \tab 1 \tab XXXX \cr
   AA \tab SIN \tab 1 \tab 2 \tab XXXX \cr
   AA \tab SIN \tab 1 \tab 3 \tab XXXX \cr
   AA \tab SIN \tab 2 \tab 1 \tab XXXX \cr
   AA \tab SIN \tab 2 \tab 2 \tab XXXX \cr
}
}
