% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-cluster.r
\name{clustEM}
\alias{clustEM}
\alias{getclustclass.clustenaem}
\alias{getclustmeans.clustenaem}
\title{Clusteriza Dado Por Mistura Gaussiana}
\usage{
clustEM(compact, nc, ...)

\method{getclustclass}{clustenaem}(clust)

\method{getclustmeans}{clustenaem}(clust)
}
\arguments{
\item{compact}{objeto \code{compactcen} contendo cenarios compactados para clusterizar}

\item{nc}{numero de clusters}

\item{...}{demais parametros passados a funcao \code{\link[mclust]{Mclust}}}

\item{clust}{objeto da classe \code{clustenaem}}
}
\value{
objeto da classe \code{clustenaem}: uma lista de um elemento contendo a saida da
clusterizacao
}
\description{
Wrapper da funcao \code{\link[mclust]{Mclust}} para uso neste pacote
}
