% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustena.r
\docType{package}
\name{clustena}
\alias{clustena}
\title{Pacote \code{clustena}}
\description{
Selecao De Cenarios De ENA
}
\details{
Este pacote prove funcoes para selecao de cenarios de Energia Natural Afluente, independentemente
da metodologia utilizada para gera-los. Alem dos metodos para selecao de cenarios sao fornecidas
ferramentas para visualizacao da escolha.
}
