% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-cluster.r
\name{clusthierarq}
\alias{clusthierarq}
\alias{getclustclass.hclust_aug}
\alias{getclustmeans.hclust_aug}
\title{Clusteriza Dado Por Metodo Aglomerativo Hierarquico}
\usage{
clusthierarq(compact, nc, distfun = dist, ...)

\method{getclustclass}{hclust_aug}(clust)

\method{getclustmeans}{hclust_aug}(clust)
}
\arguments{
\item{compact}{objeto \code{compactcen} contendo cenarios compactados para clusterizar}

\item{nc}{numero de clusters}

\item{distfun}{uma funcao que receba a matriz de dados e retorne uma medida de dissimilaridade.
Por padrao \code{\link[stats]{dist}}}

\item{...}{demais parametros passados a funcao \code{\link[stats]{hclust}}}

\item{clust}{objeto da classe \code{hclust_aug}}
}
\value{
objeto \code{hclust_aug}, que e um objeto \code{hclust} com dois atributos extras: as
classificacoes do dado cortando o dendograma em \code{nc} clusters os respectivos centroides.
Para mais detalhes acerca da classe \code{hclust}, veja \code{\link[stats]{hclust}}
}
\description{
Wrapper da funcao \code{\link[stats]{hclust}} para uso neste pacote
}
