% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-cluster.r
\name{clustkmedoids}
\alias{clustkmedoids}
\alias{getclustclass.pam}
\alias{getclustmeans.pam}
\title{Clusteriza Dado Por Kmedoides}
\usage{
clustkmedoids(compact, nc, nstart = 30, ...)

\method{getclustclass}{pam}(clust)

\method{getclustmeans}{pam}(clust)
}
\arguments{
\item{compact}{objeto \code{compactcen} contendo cenarios compactados para clusterizar}

\item{nc}{numero de clusters}

\item{nstart}{numero de sementes para testar o kmeans}

\item{...}{demais parametros passados a funcao \code{\link[cluster]{pam}} exceto \code{nstart}}

\item{clust}{objeto da classe \code{pam}}
}
\value{
objeto \code{kmeans} -- Veja \code{\link[cluster]{pam}} para mais detalhes
}
\description{
Wrapper da funcao \code{\link[cluster]{pam}} para uso neste pacote
}
