% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactacao.r
\name{compact_funs}
\alias{compact_funs}
\title{Funcoes de compactacao}
\description{
Wrappers de metodos de compactacao para uso nos processos desse pacote
}
\details{
Durante a execucao das funcoes de selecao e/ou extracao de cenarios, o objeto \code{cenariosena}
passa por uma reducao de dimensionalidade e posterior clusterizacao. As funcoes para compactacao
do dado podem implementar qualquer metodo, mas devem respeitar alguns padroes de argumentos de
entrada e caracteristicas na saida.

Quanto as entradas, qualquer funcao implementada com este proposito deve receber na primeira
posicao o argumento \code{cenarios}, um objeto da classe \code{cenariosena}. Outros argumentos
especificos a cada metodo de compactacao podem seguir o primeiro, com qualquer nome (estes
serao passados a chamada via o argumento \code{compact_args} das funcoes de selecao/extracao).

Os objetos retornados pelas funcoes de compactacao devem ser da classe \code{compactcen}. A
geracao destes objetos e feita pela funcao \code{\link{new_compactcen}}. Essencialmente deve
receber um datatable com estrutura igual a de objetos \code{cenariosena}, exceto pela coluna
\code{data} que muda para \code{ind}, indicando o indice da variavel compactada para cada
cenario. Consulte a pagina de ajuda do construtor para mais detalhes.

Atualmente o pacote fornece duas opcoes:

\itemize{
\item{\code{\link{PCAena}}}
\item{\code{\link{acumulaena}}}
}

As paginas de help de cada uma das funcoes contem detalhes a respeito dos argumentos de cada uma
(que podem ser passados as funcoes de selecao atraves de \code{compact_args} naquelas chamadas).
}
\seealso{
Funcoes \code{\link{PCAena}} e \code{\link{acumulaena}} para compactacao de dados
}
