% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-cluster.r
\name{getclust}
\alias{getclust}
\alias{getclustmeans}
\alias{getclustmeans.default}
\alias{getclustclass}
\alias{getclustclass.default}
\title{Genericas Para Classes Cluster}
\usage{
getclustmeans(clust)

\method{getclustmeans}{default}(clust)

getclustclass(clust)

\method{getclustclass}{default}(clust)
}
\arguments{
\item{clust}{objeto com classe \code{clustena} e subclasse especifica do metodo}
}
\value{
Para \code{getclustmeans}, os centroides; para \code{getclustclass} as classificacoes
}
\description{
Funcoes genericas para extracao das classes ajustadas e centroides de cada classe
}
