% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactacao.r
\name{invtransfpca}
\alias{invtransfpca}
\title{Funcao Inversa Do PCA}
\usage{
invtransfpca(pca, importance)
}
\arguments{
\item{pca}{objeto da classe \code{prcomp}}

\item{importance}{um inteiro indicando a ultima componente principal a manter}
}
\value{
funcao que recebe um vetor no espaco reduzido e retorna projecao no espaco original
}
\description{
Retorna uma funcao que recebe vetores no espaco reduzido e projeta de volta no espaco original
}
