% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-cenariosena.r
\name{learqcenarios}
\alias{learqcenarios}
\title{Leitura Do Arquivo De Cenarios}
\usage{
learqcenarios(arq, pat_data = "\%m/\%Y")
}
\arguments{
\item{arq}{caminho do arquivo a ser lido com os cenarios. Formato deve ser compativel com o
descrito em Detalhes}

\item{pat_data}{string indicando o formato de data dos nomes das colunas de cenario}
}
\value{
objeto da classe \code{cenariosena} contendo dados do arquivo lido
}
\description{
Funcao auxiliar que le arquivo texto e formata como objeto da classe \code{cenariosena}
}
\details{
Espera-se um arquivo com formato especifico. As tres primeiras colunas devem conter o indice do
cenario, bacia e ano de referencia, respectivamente. As demais colunas devem corresponder aos
passos de tempo para os quais os cenarios foram gerados. Por exemplo, um arquivo contendo cinco
cenarios gerados para 2022-01, 2022-02 e 2022-03 teria a forma\tabular{cccccc}{
   cenario \tab bacia \tab anoref \tab 1/2022 \tab 2/2022 \tab 3/2022 \cr
   1 \tab SIN \tab AA \tab XXXX \tab XXXX \tab XXXX \cr
   2 \tab SIN \tab AA \tab XXXX \tab XXXX \tab XXXX \cr
   3 \tab SIN \tab AA \tab XXXX \tab XXXX \tab XXXX \cr
   4 \tab SIN \tab AA \tab XXXX \tab XXXX \tab XXXX \cr
   5 \tab SIN \tab AA \tab XXXX \tab XXXX \tab XXXX \cr
}


XXXX   e um placeholder para os valores de ena e AA um placeholder para os valores de ano de
referencia. Nao existe um padrao para os anos, podendo ser qualquer tipo de dado (i.e. "A1",
"ano_1", 2022 e etc).

O nome das colunas de cenarios nao precisa seguir esta formato, sendo possivel informar o padrao
de nome pelo argumento \code{pat_data}, que deve corresponder a uma string de formato de data.
Neste caso, \code{pat_data = "\%m/$Y"}
}
