% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-compactcen.r
\name{new_compactcen}
\alias{new_compactcen}
\title{Construtor De \code{compactcen}}
\usage{
new_compactcen(compact, metodo, invfunc)
}
\arguments{
\item{compact}{data.table contendo a informacao de cenarios compactados. Ver Detalhes}

\item{metodo}{string indicando o nome da funcao utilizada para compactacao}

\item{infunc}{caso exista, funcao para transformar vetores no espaco reduzido de volta ao
original}
}
\value{
Objeto da classe \code{compactcen}, uma lista de um emento chamado compact contendo o
argumento \code{compact}

Adicionalmente contem os atributos

\itemize{
\item{\code{metodo}: }{string do nome da funcao chamada para compactacao}
\item{\code{teminv}: }{booleano indicando se a compactacao possui inversa}
\item{\code{invfunc}: }{caso tenha inversa, a funcao que recebe vetores no espaco compactado e
retorna no espaco original}
}
}
\description{
Construtor interno de objetos com cenarios compactados
}
\details{
O argumento \code{compact} deve seguir um padrao de formato:

\describe{
\item{\code{anoref}}{ano de referencia para geracao do cenario}
\item{\code{bacia}}{bacia a qual a ENA corresponde}
\item{\code{cenario}}{indice do cenario com respeito a bacia e ano de referencia}
\item{\code{ind}}{indice do elemento no vetor de dimensao reduzida}
\item{\code{ena}}{valor de energia afluente}
}
}
