% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactacao.r
\name{PCAena}
\alias{PCAena}
\title{Compactacao Por PCA}
\usage{
PCAena(cenarios, vartot = 0.8)
}
\arguments{
\item{cenarios}{objeto da classe \code{cenariosena} contendo apenas um ano de referencia}

\item{vartot}{percentual em formato decimal de variacao total mininima}
}
\value{
objeto da classe \code{compactcen} contendo o dado em dimensao reduzida. Este e uma lista
de um elemento nomeado \code{"compact"}, um \code{data.table} com as colunas

\describe{
\item{\code{anoref}}{ano de referencia para geracao do cenario}
\item{\code{bacia}}{bacia a qual a ENA corresponde}
\item{\code{cenario}}{indice do cenario com respeito a bacia e ano de referencia}
\item{\code{ind}}{indice do elemento no vetor de dimensao reduzida}
\item{\code{ena}}{valor de energia afluente}
}

Adicionalmente tem os atributos

\itemize{
\item{\code{metodo}: }{string do nome da funcao chamada para compactacao}
\item{\code{teminv}: }{booleano indicando se a compactacao possui inversa}
\item{\code{invfunc}: }{caso tenha inversa, a funcao que recebe vetores no espaco compactado e
retorna no espaco original}
}
}
\description{
Reduz dimensionalidade dos cenarios via PCA mantendo um minimo de variacao total
}
\details{
A matriz de dados considerada para reducao de dimensionalidade tem em cada linha o cenario e,
em cada coluna, os passos a frente de simulacao por bacia. Desta forma, cem cenarios 15 passos a
frente de duas bacias corresponde a uma matriz de dados 100 por 30. Serao selecionadas as n
primeiras componentes principais que representem no minimo \code{vartot} por cento da variacao
total. Deve ser observado que o dado e normalizado para media zero e variancia um antes da
compactacao.
}
\examples{

# usando o dado exemplo do pacote

# a compactacao deve ser usada para cada ano de referencia individualmente
cens <- cenariosdummy["A1"]

# Compactando apenas os cenarios relativos ao SIN completo
cens_compact <- PCAena(cens[, "SIN"])
\dontrun{
plot(cens_compact)
}

}
\seealso{
\code{\link{plot.compactcen}} para visualizacao das compactacoes
}
