% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizacao.r
\name{plot.cenariosena}
\alias{plot.cenariosena}
\title{Visualizacao De \code{cenariosena}}
\usage{
\method{plot}{cenariosena}(x, ...)
}
\arguments{
\item{x}{objeto da classe \code{cenariosena}}

\item{...}{nao possui funcao -- existe apenas para consistencia com a generica}
}
\description{
Plot em grade de objetos \code{cenariosena}
}
