% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizacao.r
\name{plot.compactcen}
\alias{plot.compactcen}
\title{Visualizacao De \code{compactcen}}
\usage{
\method{plot}{compactcen}(x, clusters, print = TRUE, ...)
}
\arguments{
\item{x}{objeto da classe \code{compactcen} gerado por uma das funcoes de compactacao}

\item{clusters}{argumento opcional, objeto retornado por uma das funcoes de clusterizacao}

\item{print}{booleano indicando se o plot deve ser exibido ou retornado invisivelmente}

\item{...}{nao possui utilidade, existe apenas para consistencia com a generica}
}
\value{
O plot das variaveis compactadas. Isto sera um objeto \code{ggplot} caso a compactacao
tenha sido feita em uma ou duas dimensoes ou um objeto \code{plot_ly} caso em tres.
}
\description{
Plot do dado compactado em ate tres dimensoes, possivelmente colorido por clusters estimados
}
\details{
Esta funcao permite a visualizacao da compactacao de cenarios e resultado da clusterizacao. Deve
ser notado que sua funcionalidade esta prevista apenas para compactacoes em ate tres dimensoes e
retornara um erro caso seja passado \code{x} com quatro ou mais.
}
