% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.r
\name{selecporcluster}
\alias{selecporcluster}
\title{Selecao Ou Extracao De Cenarios Por Clusterizacao}
\usage{
selecporcluster(
  cenarios,
  nc,
  clust_fun = clustkmeans,
  ...,
  transforma = FALSE,
  compact_fun = PCAena,
  compact_args = list()
)
}
\arguments{
\item{cenarios}{objeto da classe \code{cenariosena} do qual selecionar cenarios representativos
contendo apenas uma bacia e ano de referencia}

\item{nc}{numero de clusters}

\item{clust_fun}{funcao para clusterizacao. Veja \code{\link{clust_funs}} para mais detalhes}

\item{...}{demais parametros passados a \code{clust_fun}}

\item{transforma}{booleano indicando se devem ser retornados indices dos cenarios mais proximos
do centroide de cada cluster (\code{FALSE}) ou se o cetroide deve ser transformado de volta
para a escala original (\code{TRUE}). Ver Detalhes}

\item{compact_fun}{funcao para compactacao dos cenarios. Ver \code{\link{compact_funs}}}

\item{compact_args}{lista nomeada informando argumentos (alem do primeiro) a \code{compact_fun}}
}
\value{
lista contendo vetor de inteiros indicando indices dos cenarios selecionados, resultado
da compactacao e resultado da clusterizacao
}
\description{
Seleciona ou extrai cenarios baseado nos centroides da clusterizacao de cenarios
}
\details{
O procedimento desta funcao e aplicar a reducao de dimensionalidade em \code{cenarios} e
clusterizar as variaveis no espaco reduzido em \code{nc} clusters.

A compactacao pode ser feita por qualquer funcao que retorne objetos da classe \code{compactcen},
informada atraves do argumento \code{compact_fun}. A funcao sera chamada com \code{cenarios} como
primeiro arumento; \code{compact_args} permite informar demais parametros a serem utilizados por
\code{compact_fun}, na forma de uma lista nomeada.

A clusterizacao em si e realizada por uma chamada a funcao passada pelo argumento
\code{clust_fun}, uma funcao que deve receber como primeiro argumento um objeto \code{cenarios} e
como segundo argumento o numero de clusters desejado \code{nc}, e retornar um objeto com os
metodos \code{getclustclass} e \code{getclustmeans}, para retornar as classificacoes de cada
cenario e as medias dos centroides. A funcao passada sera usada numa chamada da forma
\code{clust_fun(cenarios, nc, ...)}. \code{compact_fun} e \code{compact_args}.

Por fim, \code{transforma} controla o que sera retornado. Se \code{FALSE}, a funcao entrega
apenas os indices dos cenarios mais proximos, por distancia euclidiana, dos centroides de cada
cluster. Se \code{TRUE}, o centroide de cada cluster (no espaco reduzido) sera transformado de
volta para o espaco original. Isso so e possivel caso a compactacao utilizada permita este tipo
de operacao: por exemplo, a reducao de dimensionalidade por acumulo de enas parciais e uma
transformacao que nao permite retornar para a escala original. Outras como PCA e autoencoder sim.
}
