% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.r
\name{selecporquantil}
\alias{selecporquantil}
\title{Selecao De Cenarios Por Quantil}
\usage{
selecporquantil(
  cenarios,
  quantis = c(0.25, 0.5, 0.75),
  compact_fun = acumulaena,
  compact_args = list()
)
}
\arguments{
\item{cenarios}{objeto da classe \code{cenariosena} do qual selecionar cenarios representativos
contendo apenas uma bacia e ano de referencia}

\item{quantis}{vetor de quantis para selecionar}

\item{compact_fun}{funcao para compactacao dos cenarios. Ver \code{\link{compact_funs}}}

\item{compact_args}{lista nomeada informando argumentos (alem do primeiro) a \code{compact_fun}}
}
\value{
lista contendo vetor de inteiros indicando indices dos cenarios selecionados e resultado
da compactacao
}
\description{
Seleciona cenarios representativos buscando aqueles mais proximos de quantis especificados
}
\details{
A selecao por quantis requer que cada cenario seja reduzido a apenas uma dimensao (de modo que
seja possivel calcular quantis unicos). A compactacao pode ser feita por qualquer funcao que
retorne objetos da classe \code{compactcen}, informada atraves do argumento \code{compact_fun}. A
funcao sera chamada com \code{cenarios} como primeiro arumento; \code{compact_args} permite
informar demais parametros a serem utilizados por \code{compact_fun}, na forma de uma lista
nomeada.

O argumento cenarios deve ser um objeto da classe \code{cenariosena} contendo apenas um ano de
referencia e bacia. \code{quantis} deve ser um vetor indicando quais quantis estao sendo
buscados. Por construcao e impossivel extrair exatamente estes quantis, de modo que serao
retornados os indices dos cenarios associados aos quantis mais proximos.
}
