% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactacao.r
\name{stepcumsum}
\alias{stepcumsum}
\title{Soma Acumulada Por Degraus}
\usage{
stepcumsum(x, qbr)
}
\arguments{
\item{x}{vetor do qual calcular soma}

\item{qbr}{escalar ou vetor de inteiros indicando quantas partes ou onde separar as partes}
}
\value{
vetor de ENA acumulada por partes
}
\description{
Calcula a soma acumulada de um vetor em partes
}
