% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-cenariosena.r
\name{[.cenariosena}
\alias{[.cenariosena}
\title{Subset De \code{cenariosena}}
\usage{
\method{[}{cenariosena}(x, i, j, k, l, ...)
}
\arguments{
\item{x}{objeto da classe \code{cenariosena}}

\item{i}{escalar ou vetor de anos de referencia para extrair do dado}

\item{j}{escalar ou vetor de bacias para extrair do dado}

\item{k}{escalar ou vetor de indices de cenarios para extrair do dado}

\item{l}{escalar ou vetor de datas para extrair do dado}

\item{...}{nao possui uso -- existe apenas para consistencia com a generica}
}
\value{
objecto \code{cenariosena} contendo apenas os valores especificados
}
\description{
Extrai subsets de objetos \code{cenariosena} e adequa os atributios
}
\examples{

# utilizando o dado exemplo do pacote

# subset apenas do ano de referencia "A1"
cens <- cenariosdummy["A1"]

# subset apenas da regiao "SUL"
cens <- cenariosdummy[, "SUL"]

# subset dos cenarios 5, 10, 20
cens <- cenariosdummy[, , c(5, 10, 20)]

# datas simuladas 2022-11-01/2023-04-01
cens <- cenariosdummy[, , , seq.Date(as.Date("2022-11-01"), as.Date("2023-04-01"), by = "month")]

# ano "A2", NE e N, cenarios 10, 11, 12 nas datas 2022-09-01/2023-02-01
cens <- cenariosdummy[
    "A2",
    c("NE", "N"),
    c(10, 11, 12),
    seq.Date(as.Date("2022-09-01"), as.Date("2023-02-01"), by = "month")
]

}
