% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactacao.r
\name{acumulaena}
\alias{acumulaena}
\title{Calcula ENA Acumulada}
\usage{
acumulaena(cenarios, quebras = 1L)
}
\arguments{
\item{cenarios}{objeto da classe \code{cenariosena} contendo apenas uma bacia e ano de referencia}

\item{quebras}{ou um inteiro indicando em quantas partes iguais separar o dado ou um vetor de
inteiros indicando as posicoes nas quais separar}
}
\value{
objeto da classe \code{compactcen} contendo o dado em dimensao reduzida
}
\description{
Reducao de dimensao por ENA acumulada, possivelmente em partes crescentes
}
\details{
Esta funcao reduz a dimensao de um vetor de cenarios de ENA calculando a soma acumulada do vetor,
possivelmente em partes. Se \code{quebras = 3L}, por exemplo, a funcao quebra o cenario em tres
partes iguais e calcula a soma de cada um. O vetor em dimensao reduzida e, entao, a soma
acumulada destas partes.

Por exemplo, para cenarios de um ano, indo de janeiro a dezembro, o uso de \code{quebras = 3L}
representa o calculo da ENA acumulada ate abril, agosto e dezembro (por construcao um vetor de
valores crescentes).
}
