% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-cluster.r
\name{clust_funs}
\alias{clust_funs}
\title{Funcoes De Clusterizacao}
\description{
Wrappers de metodos de clusterizacao para uso nos processos do pacote
}
\details{
Durante a execucao das funcoes de selecao e/ou extracao de cenarios, o objeto \code{cenariosena}
passa por uma reducao de dimensionalidade e posterior clusterizacao. As funcoes para clusterizar
o dado podem implementar qualquer metodo, mas devem respeitar alguns padroes de argumentos de
entrada e caracteristicas na saida.

Quanto as entradas, todas as funcoes de clusterizacao precisam receber o primeiro argumento
chamado \code{compact}, correspondente a um objeto \code{compactcen} observacoes a serem
clusterizadas. Alem dele, o segundo argumento deve ser \code{nc}, recebendo um inteiro indicando
o numero de clustes desejados. Os demais argumentos sao livres, porem independentemente deles
deve existir \code{...} para consistencia com as demais funcoes.

Nao ha restricoes ou especificacoes para o objeto de saida, porem devem ser definidos metodos das
genericas \code{getclustclass} e \code{getclustmeans}, que extraem do objeto de saida as
classificacoes de cada obsevacao e centroides dos clusters respectivamente, para a classe dos
objetos retornados. Por exemplo, \code{clustkmeans} retorna um objeto com classe \code{kmeans},
de modo que ha implementadas no pacote \code{getclustmeans.kmeans} e \code{getclustclass.kmeans},
que recebem objetos dessa classe e extraem centroides e classficacoes respectivamente.

Atualmente o pacote fornece as opcoes:

\itemize{
\item{\code{\link{clustkmeans}}}
\item{\code{\link{clustkmedoids}}}
\item{\code{\link{clustEM}}}
\item{\code{\link{clusthierarq}}}
}

As paginas de help de cada uma das funcoes contem detalhes a respeito dos argumentos de cada uma
(que podem ser passados as funcoes de selecao atraves de \code{...} naquelas chamadas).
}
