% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-cluster.r
\name{clustEM}
\alias{clustEM}
\alias{getclustclass.Mclust}
\alias{getclustmeans.Mclust}
\title{Clusteriza Dado Por Mistura Gaussiana}
\usage{
clustEM(compact, nc, ...)

\method{getclustclass}{Mclust}(clust)

\method{getclustmeans}{Mclust}(clust)
}
\arguments{
\item{compact}{objeto \code{compactcen} contendo cenarios compactados para clusterizar}

\item{nc}{numero de clusters}

\item{...}{demais parametros passados a funcao \code{\link[mclust]{Mclust}}}

\item{clust}{objeto da classe \code{Mclust}}
}
\value{
objeto \code{Mclust} -- Veja \code{\link[mclust]{Mclust}} para mais detalhes
}
\description{
Wrapper da funcao \code{\link[mclust]{Mclust}} para uso neste pacote
}
