% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-cluster.r
\name{clustkmeans}
\alias{clustkmeans}
\alias{getclustclass.clustenakmeans}
\alias{getclustmeans.clustenakmeans}
\title{Clusteriza Dado Por Kmeans}
\usage{
clustkmeans(compact, nc, nstart = 30, ...)

\method{getclustclass}{clustenakmeans}(clust)

\method{getclustmeans}{clustenakmeans}(clust)
}
\arguments{
\item{compact}{objeto \code{compactcen} contendo cenarios compactados para clusterizar}

\item{nc}{numero de clusters}

\item{nstart}{numero de sementes para testar o kmeans}

\item{...}{demais parametros passados a funcao \code{\link[stats]{kmeans}} exceto \code{nstart}}

\item{clust}{objeto da classe \code{clustenakmeans}}
}
\value{
objeto da classe \code{clustenakmeans}: uma lista de um elemento contendo a saida da
clusterizacao
}
\description{
Wrapper da funcao \code{\link[stats]{kmeans}} para uso neste pacote
}
