% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactacao.r
\name{PCAena}
\alias{PCAena}
\title{Compactacao Por PCA}
\usage{
PCAena(cenarios, vartot = 0.8)
}
\arguments{
\item{cenarios}{objeto da classe \code{cenariosena} contendo apenas uma bacia e ano de referencia}

\item{vartot}{percentual em formato decimal de variacao total mininima}
}
\value{
objeto da classe \code{compactcen} contendo o dado em dimensao reduzida
}
\description{
Reduz dimensionalidade dos cenarios via PCA mantendo um minimo de variacao total
}
\details{
A matriz de dados considerada para reducao de dimensionalidade tem em cada linha o cenario e,
em cada coluna, os passos a frente de simulacao. Serao selecionadas as n primeiras componentes
principais que representem no minimo \code{vartot} por cento da variacao total.
}
