% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizacao.r
\name{plot.cenariosena}
\alias{plot.cenariosena}
\title{Visualizacao De \code{cenariosena}}
\usage{
\method{plot}{cenariosena}(x, ..., print = TRUE)
}
\arguments{
\item{x}{objeto da classe \code{cenariosena}}

\item{...}{objetos \code{cenariosena} opcionais com cenarios a serem plotados por cima}

\item{print}{booleano indicando se o plot deve ser exibido ou retornado invisivelmente}
}
\description{
Plot em grade de objetos \code{cenariosena}
}
\details{
O argumento \code{...} permite que sejam passados mais outros objetos \code{cenariosena} contendo
um numero menor de cenarios selecionados para plot por cima do completo. A intencao de uso deste
argumento e passar \code{cenariosena} para cada ano dos quais se tiram cenarios, fazendo a
selecao do mesmo numero de cenarios a cada ano de referencia. Assim, se existem tres anos de
referencia, \code{...} corresponde a tres objetos \code{cenariosena} contendo cada um um unico
ano de referencia e o mesmo numero N de cenarios selecionados.
}
