% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-cenariosena.r
\name{[.cenariosena}
\alias{[.cenariosena}
\title{Subset De \code{cenariosena}}
\usage{
\method{[}{cenariosena}(x, i, j, k, l, ...)
}
\arguments{
\item{x}{objeto da classe \code{cenariosena}}

\item{i}{escalar ou vetor de anos de referencia para extrair do dado}

\item{j}{escalar ou vetor de bacias para extrair do dado}

\item{k}{escalar ou vetor de indices de cenarios para extrair do dado}

\item{l}{escalar ou vetor de datas para extrair do dado}

\item{...}{nao possui uso -- existe apenas para consistencia com a generica}
}
\value{
objecto \code{cenariosena} contendo apenas os valores especificados
}
\description{
Extrai subsets de objetos \code{cenariosena} e adequa os atributios
}
